--- src/Stratis.Bitcoin.Features.Consensus/CoinViews/Coindb/BaseCoindb.cs
@@ -1,214 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using Microsoft.Extensions.Logging;
-using NBitcoin;
-using Stratis.Bitcoin.Interfaces;
-using Stratis.Bitcoin.Utilities;
-
-namespace Stratis.Bitcoin.Features.Consensus.CoinViews
-{
-    /// <summary>
-    /// A base class for <see cref="LevelDbCoindb"/> and implementations related to other database types.
-    /// </summary>
-    public class BaseCoindb
-    {
-        protected static readonly byte coinsTable = 1;
-        protected static readonly byte blockTable = 2;
-        protected static readonly byte rewindTable = 3;
-        protected static readonly byte stakeTable = 4;
-        protected static readonly byte balanceTable = 5;
-        protected static readonly byte balanceAdjustmentTable = 6;
-
-        /// <summary>Database key under which the block hash of the coin view's current tip is stored.</summary>
-        protected static readonly byte[] blockHashKey = new byte[0];
-
-        /// <summary>Database key under which the block hash of the coin view's last indexed tip is stored.</summary>
-        protected static readonly byte[] blockIndexedHashKey = new byte[1];
-
-        /// <summary>Instance logger.</summary>
-        protected ILogger logger;
-
-        public bool BalanceIndexingEnabled { get; protected set; }
-
-        /// <summary>Access to dBreeze database.</summary>
-        protected IDb coinDb;
-
-        /// <summary>Hash of the block which is currently the tip of the coinview.</summary>
-        protected HashHeightPair persistedCoinviewTip;
-
-        private readonly IScriptAddressReader scriptAddressReader;
-
-        private readonly Network network;
-
-        public BaseCoindb(Network network, IScriptAddressReader scriptAddressReader)
-        {
-            this.network = network;
-            this.scriptAddressReader = scriptAddressReader;
-        }
-
-        public IEnumerable<(uint height, long satoshis)> GetBalance(TxDestination txDestination)
-        {
-            long balance = 0;
-            {
-                byte[] row = this.coinDb.Get(balanceTable, txDestination.ToBytes());
-                balance = (row == null) ? 0 : BitConverter.ToInt64(row);
-            }
-
-            foreach ((byte[] key, byte[] value) in this.coinDb.GetAll(balanceAdjustmentTable, ascending: false, 
-                lastKey: txDestination.ToBytes().Concat(BitConverter.GetBytes(this.persistedCoinviewTip.Height + 1).Reverse()).ToArray(), 
-                includeLastKey: false,
-                firstKey: txDestination.ToBytes(),
-                includeFirstKey: false))
-            {
-                yield return (BitConverter.ToUInt32(key.Reverse().ToArray()), balance);
-                balance -= BitConverter.ToInt64(value);
-            }
-
-            yield return (0, balance);
-        }
-
-        /// <summary>
-        /// The 'skipMissing' flag allows us to rewind coind db's that have incomplete balance information.
-        /// </summary>
-        protected void AdjustBalance(IDbBatch batch, Dictionary<TxDestination, Dictionary<uint, long>> balanceUpdates)
-        {
-            foreach ((TxDestination txDestination, Dictionary<uint, long> balanceAdjustments) in balanceUpdates)
-            {
-                long totalAdjustment = 0;
-
-                foreach (uint height in balanceAdjustments.Keys.OrderBy(k => k))
-                {
-                    var key = txDestination.ToBytes().Concat(BitConverter.GetBytes(height).Reverse()).ToArray();
-                    byte[] row = this.coinDb.Get(balanceAdjustmentTable, key);
-                    long balance = ((row == null) ? 0 : BitConverter.ToInt64(row)) + balanceAdjustments[height];
-                    batch.Put(balanceAdjustmentTable, key, BitConverter.GetBytes(balance));
-
-                    totalAdjustment += balance;
-                }
-
-                {
-                    var key = txDestination.ToBytes();
-                    byte[] row = this.coinDb.Get(balanceTable, key);
-                    long balance = ((row == null) ? 0 : BitConverter.ToInt64(row)) + totalAdjustment;
-                    batch.Put(balanceTable, key, BitConverter.GetBytes(balance));
-                }
-            }
-        }
-
-        protected void Update(Dictionary<TxDestination, Dictionary<uint, long>> balanceAdjustments, Script scriptPubKey, uint height, long change)
-        {
-            if (scriptPubKey.Length == 0 || change == 0)
-                return;
-
-            foreach (TxDestination txDestination in this.scriptAddressReader.GetDestinationFromScriptPubKey(this.network, scriptPubKey))
-            {
-                if (!balanceAdjustments.TryGetValue(txDestination, out Dictionary<uint, long> value))
-                {
-                    value = new Dictionary<uint, long>();
-                    balanceAdjustments[txDestination] = value;
-                }
-
-                if (!value.TryGetValue(height, out long balance))
-                    balance = change;
-                else
-                    balance += change;
-
-                value[height] = balance;
-            }
-        }
-
-        public HashHeightPair GetTipHash()
-        {
-            if (this.persistedCoinviewTip == null)
-            {
-                var row = this.coinDb.Get(blockTable, blockHashKey);
-                if (row != null)
-                {
-                    this.persistedCoinviewTip = new HashHeightPair();
-                    this.persistedCoinviewTip.FromBytes(row);
-                }
-            }
-
-            return this.persistedCoinviewTip;
-        }
-
-        protected HashHeightPair GetIndexedTipHash()
-        {
-            var row = this.coinDb.Get(blockTable, blockIndexedHashKey);
-            if (row != null)
-            {
-                var tip = new HashHeightPair();
-                tip.FromBytes(row);
-                return tip;
-            }
-
-            return null;
-        }
-
-        protected void SetBlockHash(IDbBatch batch, HashHeightPair nextBlockHash, bool forceUpdateIndexedHeight = false)
-        {
-            this.persistedCoinviewTip = nextBlockHash;
-            batch.Put(blockTable, blockHashKey, nextBlockHash.ToBytes());
-            if (this.BalanceIndexingEnabled || forceUpdateIndexedHeight)
-                batch.Put(blockTable, blockIndexedHashKey, nextBlockHash.ToBytes());
-        }
-
-        protected virtual HashHeightPair RewindInternal(int startHeight, HashHeightPair target)
-        {
-            throw new NotImplementedException();
-        }
-
-        protected void EnsureCoinDatabaseIntegrity(ChainedHeader chainTip)
-        {
-            this.logger.LogInformation("Checking coin database integrity...");
-
-            HashHeightPair maxHeight = new HashHeightPair(chainTip);
-
-            // If the balance table is empty then rebuild the coin db.
-            if (this.BalanceIndexingEnabled)
-            {
-                HashHeightPair indexedTipHash = this.GetIndexedTipHash();
-                if (indexedTipHash == null)
-                {
-                    this.logger.LogInformation($"Rebuilding coin database to include balance information.");
-                    this.coinDb.Clear();
-                    return;
-                }
-
-                if (indexedTipHash.Height < chainTip.Height)
-                {
-                    this.logger.LogInformation($"Rewinding the coin database to include missing balance information.");
-                    maxHeight = indexedTipHash;
-                }
-            }
-
-            var tipHash = GetTipHash();
-
-            var heightToCheck = chainTip.Height;
-
-            if (heightToCheck > tipHash.Height)
-                heightToCheck = tipHash.Height;
-
-            // Find the height up to where rewind data is stored above chain tip.
-            do
-            {
-                heightToCheck += 1;
-
-                byte[] row = this.coinDb.Get(rewindTable, BitConverter.GetBytes(heightToCheck).Reverse().ToArray());
-                if (row == null)
-                    break;
-            } while (true);
-
-            for (int height = heightToCheck - 1; height > maxHeight.Height;)
-            {
-                this.logger.LogInformation($"Fixing coin database, deleting rewind data at height {height} above tip '{chainTip}'.");
-
-                // Do a batch of rewinding.
-                height = RewindInternal(height, maxHeight).Height;
-            }
-
-            this.logger.LogInformation("Coin database integrity good.");
-        }
-    }
-}
