+++ src/Stratis.Bitcoin.IntegrationTests/NodeContext.cs
@@ -28,7 +28,7 @@ namespace Stratis.Bitcoin.IntegrationTests
             this.FolderName = TestBase.CreateTestDir(caller, name);
             var dateTimeProvider = new DateTimeProvider();
             var serializer = new DBreezeSerializer(this.Network.Consensus.ConsensusFactory);
-            this.Coindb = new LevelDbCoindb(network, this.FolderName, dateTimeProvider, new NodeStats(dateTimeProvider, NodeSettings.Default(network), new Mock<IVersionProvider>().Object), serializer, new ScriptAddressReader());
+            this.Coindb = new Coindb<LevelDb>(network, new DataFolder(this.FolderName), dateTimeProvider, new NodeStats(dateTimeProvider, NodeSettings.Default(network), new Mock<IVersionProvider>().Object), serializer, new ScriptAddressReader());
             this.Coindb.Initialize(new ChainedHeader(network.GetGenesis().Header, network.GenesisHash, 0), false);
             this.cleanList = new List<IDisposable> { (IDisposable)this.Coindb };
         }
@@ -66,7 +66,7 @@ namespace Stratis.Bitcoin.IntegrationTests
             this.cleanList.Remove((IDisposable)this.Coindb);
             var dateTimeProvider = new DateTimeProvider();
             var serializer = new DBreezeSerializer(this.Network.Consensus.ConsensusFactory);
-            this.Coindb = new LevelDbCoindb(this.Network, this.FolderName, dateTimeProvider, new NodeStats(dateTimeProvider, NodeSettings.Default(this.Network), new Mock<IVersionProvider>().Object), serializer, new ScriptAddressReader());
+            this.Coindb = new Coindb<LevelDb>(this.Network, new DataFolder(this.FolderName), dateTimeProvider, new NodeStats(dateTimeProvider, NodeSettings.Default(this.Network), new Mock<IVersionProvider>().Object), serializer, new ScriptAddressReader());
 
             this.Coindb.Initialize(chainTip, false);
             this.cleanList.Add((IDisposable)this.Coindb);
